Store the hash of the token returned to the client in ACCESS_TOKEN_HASH column
Token lookup happens with the hash of the access token sent by the client. Therefore when persisting the
access token we need to make sure we persist the hash of the access token returned to the client.

In the case of plain UUID access token the token persisted in the DB (ACCESS_TOKEN column) and the token returned to the client are the same.
But for JWT access tokens we perisist the jti of the JWT instead of the JWT returned to the client.