Fix app ownership update logic
According to the current logic, if the app owner we are trying to update is an invalid one, then the ownership goes to the user who calls the OAuthAdminService#updateConsumerApplication().

Instead, this fixes the logic to keep the **current app** owner to be the app owner if the user we are trying to update as the new app owner is an invalid user.